We defer the closing of the database
We previously didn't stop the timer at the end of the benchmark so it was factoring in the time it took to close the DB.
Now for every benchmark, we stop timer at end of loop, this should make benchmarks way more stable

Should close https://github.com/Viq111/kvimd/issues/13