@Viq111 2018-12-18
<details>
 <summary>Extended benchmarks results </summary>

```gf
name                           old time/op    new time/op    delta
HashDiskSetWithLoad_0.00-0.50     466ns ± 0%     541ns ± 0%   ~     (p=1.000 n=1+1)
HashDiskSetWithLoad_0.70-0.90    1.23µs ± 0%    1.25µs ± 0%   ~     (p=1.000 n=1+1)
HashDiskSetWithLoad_0.90-0.95    2.16µs ± 0%    3.22µs ± 0%   ~     (p=1.000 n=1+1)
HashDiskSetWithLoad_0.95-0.99    17.6µs ± 0%    17.9µs ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.00-0.50     227ns ± 0%     247ns ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.70-0.90     658ns ± 0%     622ns ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.90-0.95    2.01µs ± 0%    2.04µs ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.95-0.99    15.9µs ± 0%    15.8µs ± 0%   ~     (p=1.000 n=1+1)

name                           old speed      new speed      delta
HashDiskSetWithLoad_0.00-0.50  51.5MB/s ± 0%  44.3MB/s ± 0%   ~     (p=1.000 n=1+1)
HashDiskSetWithLoad_0.70-0.90  19.5MB/s ± 0%  19.2MB/s ± 0%   ~     (p=1.000 n=1+1)
HashDiskSetWithLoad_0.90-0.95  11.1MB/s ± 0%   7.5MB/s ± 0%   ~     (p=1.000 n=1+1)
HashDiskSetWithLoad_0.95-0.99  1.36MB/s ± 0%  1.34MB/s ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.00-0.50   105MB/s ± 0%    97MB/s ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.70-0.90  36.4MB/s ± 0%  38.5MB/s ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.90-0.95  11.9MB/s ± 0%  11.8MB/s ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.95-0.99  1.51MB/s ± 0%  1.52MB/s ± 0%   ~     (p=1.000 n=1+1)
```